requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
base58==0.2.5
certifi==2019.9.11
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
django-extensions==2.2.5
django-rest-swagger==2.2.0
django==2.2.7
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
openapi-codec==1.3.2
pip-tools==4.2.0
psycopg2-binary==2.8.4
pyblake2==1.1.2
python-axolotl-curve25519==0.4.1.post2
pytz==2019.3
pywaves==0.8.25.post2
requests==2.22.0
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
